+++ thrift/src/main/scala/com/twitter/finatra/thrift/filters/package.scala
@@ -1,7 +1,0 @@
-package com.twitter.finatra.thrift
-
-package object filters {
-
-  @deprecated("Use the ClientIdAcceptlistFilter", "2018-11-08")
-  type ClientIdWhitelistFilter = ClientIdAcceptlistFilter
-}