+++ thrift/src/test/scala/com/twitter/finatra/thrift/tests/ThriftServerStartupFeatureTest.scala
@@ -1,6 +1,5 @@
 package com.twitter.finatra.thrift.tests
 
-import com.twitter.finatra.thrift.modules.ClientIdAcceptlistModule
 import com.twitter.finatra.thrift.routing.ThriftRouter
 import com.twitter.finatra.thrift.tests.doeverything.controllers.DoNothingController
 import com.twitter.finatra.thrift.{EmbeddedThriftServer, ThriftServer}
@@ -12,8 +11,6 @@ class ThriftServerStartupFeatureTest extends Test {
 
   test("fail startup") {
     val server = new EmbeddedThriftServer(new ThriftServer {
-        override val modules = Seq(new ClientIdAcceptlistModule("/clients.yml"))
-
         override def configureThrift(router: ThriftRouter): Unit = {
           router.add[DoNothingController]
           router.add[DoNothingController] // can't add two services
@@ -32,8 +29,6 @@ class ThriftServerStartupFeatureTest extends Test {
 
   test("use default thrift port") {
     val server = new EmbeddedTwitterServer(new ThriftServer {
-        override val modules = Seq(new ClientIdAcceptlistModule("/clients.yml"))
-
         override def configureThrift(router: ThriftRouter): Unit = {
           router
             .add[DoNothingController]
@@ -53,8 +48,6 @@ class ThriftServerStartupFeatureTest extends Test {
 
   test("start with directly instantiated controller") {
     val server = new EmbeddedTwitterServer(new ThriftServer {
-        override val modules = Seq(new ClientIdAcceptlistModule("/clients.yml"))
-
         override def configureThrift(router: ThriftRouter): Unit = {
           router
           // directly instantiate controller and add it to the router